src/main/java/eu/erasmuswithoutpaper/course/entity/AcademicTerm.java

package eu.erasmuswithoutpaper.course.entity;

import java.io.Serializable;
import java.util.Date;
import javax.persistence.EmbeddedId;
import javax.persistence.Entity;
import javax.persistence.Temporal;
import javax.persistence.TemporalType;

@Entity
public class AcademicTerm implements Serializable{

    @EmbeddedId
    private AcademicTermId academicTermId;
    private String dispName;
    
    @Temporal(TemporalType.DATE)
    private Date startDate;
    
    @Temporal(TemporalType.DATE)
    private Date endDate;

    public AcademicTermId getAcademicTermId() {
        return academicTermId;
    }

    public void setAcademicTermId(AcademicTermId academicTermId) {
        this.academicTermId = academicTermId;
    }

    public String getDispName() {
        return dispName;
    }

    public void setDispName(String dispName) {
        this.dispName = dispName;
    }

    public Date getStartDate() {
        return startDate;
    }

    public void setStartDate(Date startDate) {
        this.startDate = startDate;
    }

    public Date getEndDate() {
        return endDate;
    }

    public void setEndDate(Date endDate) {
        this.endDate = endDate;
    }
    
}
